requirements.txt
aiohttp
beautifulsoup4
rich
brotli
stem
PySocks